+++ packages/_vue3-migration-test/.eslintrc.js
@@ -3,5 +3,8 @@ module.exports = {
   parserOptions: {
     tsconfigRootDir: __dirname,
     project: 'tsconfig.eslint.json'
+  },
+  rules: {
+    'max-len': 'off'
   }
 };