+++ packages/_vue3-migration-test/src/components/index.ts
@@ -1,4 +1,6 @@
 export * from './animations';
+export { default as TestBaseColumnPickerDropdown } from './column-picker/test-base-column-picker-dropdown.vue';
+export { default as TestBaseColumnPickerList } from './column-picker/test-base-column-picker-list.vue';
 export { default as TestBaseDropdown } from './test-base-dropdown.vue';
 export * from './scroll';
 export { default as TestBaseEventButton } from './test-base-event-button.vue';